pyproject.toml
[project]
name = "scftp"
version = "0.1.0"
description = "基于云函数的隧道代理 (Tunnel proxy based on Serverless Cloud Function)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "mitmproxy>=11.0.2",
    "orjson>=3.11.3",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
